+++ src/RPCServer.php
@@ -12,6 +12,7 @@ use Rubix\Server\Commands\PredictSample;
 use Rubix\Server\Commands\Proba;
 use Rubix\Server\Commands\ProbaSample;
 use Rubix\Server\Commands\Rank;
+use Rubix\Server\Commands\RankSample;
 use Rubix\Server\Commands\QueryModel;
 use Rubix\Server\Commands\ServerStatus;
 use Rubix\Server\Handlers\PredictHandler;
@@ -19,6 +20,7 @@ use Rubix\Server\Handlers\PredictSampleHandler;
 use Rubix\Server\Handlers\ProbaHandler;
 use Rubix\Server\Handlers\ProbaSampleHandler;
 use Rubix\Server\Handlers\RankHandler;
+use Rubix\Server\Handlers\RankSampleHandler;
 use Rubix\Server\Handlers\QueryModelHandler;
 use Rubix\Server\Handlers\ServerStatusHandler;
 use Rubix\Server\Http\Controllers\RPCController;
@@ -257,6 +259,7 @@ class RPCServer implements Server, LoggerAware
 
         if ($estimator instanceof Ranking) {
             $commands[Rank::class] = new RankHandler($estimator);
+            $commands[RankSample::class] = new RankSampleHandler($estimator);
         }
 
         return new CommandBus($commands);