+++ CHANGELOG.md
@@ -7,3 +7,4 @@
     - Added REST and RPC controllers
     - Added predict single sample command
     - Added predict probabilities of a single sample
+    - Added rank single sample command